refactor: moves pages to src folder
Moves the pages directory to `src/pages` to keep all scripts within src
