+++ __tests__/pages/index.test.jsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import IndexPage from '../../pages/index';
+import IndexPage from '../../src/pages/index';
 
 describe('<IndexPage />', () => {
   it('renders without crashing', () => {
